// src/codex/generated/v2/ModelListResponse.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Model } from "./Model.js";

export type ModelListResponse = {
  data: Array<Model>;
  /**
   * Opaque cursor to pass to the next call to continue after the last item.
   * If None, there are no more items to return.
   */
  nextCursor: string | null;
};